TCPServer - Redis RPUSH BulkQuery, Critical Bug (Connection Leak)
Fixed a critical bug where connections were being opened / closed after every message was processed. Another critical bug was found where all redis connections were never closed causing connection leak. Now KICallback has connections to both Redis and DB and these are passed into Factory and Strategies to initiate actions.

Changed the way BulkQuery behaves when talking to Redis - it used to be that using BulkQuery just meant looping and string setting all elements into keys. This is no longer the case. BulkQuery is now reserved for special operations. When BulkQuery is true, an RPUSH is done to Redis into a new/existing list for each item in the data array. If not using BulkQuery (false), will use the normal String Set operation on a key.

Updated Unit Tests to check for RPUSH commands and also check number of times a PUBLISH occurs for the same key.

Updated ProcessMessageStrategyFactory class to now accept both IConnectionMultiplexer and IDbConnection as arguments - only 1 of these will be used depending on the EDbSource type, but it was the best way of abstracting away knowing which connection object to pass in from the calling code.