Bug Fix #54
Reworked the way the throttling mechanism works - now it is a limit on the average number of requests per second. If there are more requests coming in per second than the average limit, those requests are dropped.

Updated Unit Test cases, and added new class RequestInfo that stores the throttle info.